navbar refactor and temporarily removed redirects
