+++ client/src/components/pages/Home/index.js
@@ -2,8 +2,6 @@ import React, { Component } from "react";
 import { Redirect } from "react-router-dom";
 import "./media.css";
 import "./style.css";
-
-import NavLI from "../../Navbar/loggedIn";
 import Container from "react-bootstrap/Container";
 import Row from "react-bootstrap/Row";
 import Col from "react-bootstrap/Col";
@@ -386,7 +384,6 @@ class Home extends Component {
         return (
         <div className="chat-page">
             {this.renderRedirect()}
-            <NavLI />
                 <Container className="homeContainer">
 
                     <Row className="userActiveChats">
